+++ debugger/src/ritz/jpda/jdi.clj
@@ -73,6 +73,7 @@
            option-args (.get arguments "options")
            args (str " -cp '" classpath "' clojure.main -e '" expr "'")]
        (logging/trace "jdi/launch %s" args)
+       (logging/trace "jdi/launch options %s" (string/join " " options))
        (.setValue quote-args "'")
        (.setValue main-args args)
        (.setValue option-args (string/join " " options))
@@ -277,21 +278,26 @@
   ;;  "jdi/invoke-method %s %s\nargs %s\noptions %s"
   ;;  class-or-object method (pr-str args) options)
   (logging/trace "jdi/invoke-method %s %s" method options)
-  (letfn [(invoke []
-            (let [args (java.util.ArrayList. (or args []))]
-                    (cond
-                      (instance? com.sun.jdi.ClassType class-or-object)
-                      (.invokeMethod
-                       ^ClassType class-or-object thread
-                       method args (int threading))
-                      (instance? com.sun.jdi.ObjectReference class-or-object)
-                      (.invokeMethod
-                       ^ObjectReference class-or-object thread
-                       method args (int threading)))))]
-    (if disable-exception-requests
-      (with-disabled-exception-requests [(.virtualMachine thread)]
-        (invoke))
-      (invoke))))
+  (try
+    (letfn [(invoke []
+              (let [args (java.util.ArrayList. (or args []))]
+                (cond
+                  (instance? com.sun.jdi.ClassType class-or-object)
+                  (.invokeMethod
+                   ^ClassType class-or-object thread
+                   method args (int threading))
+                  (instance? com.sun.jdi.ObjectReference class-or-object)
+                  (.invokeMethod
+                   ^ObjectReference class-or-object thread
+                   method args (int threading)))))]
+      (if disable-exception-requests
+        (with-disabled-exception-requests [(.virtualMachine thread)]
+          (invoke))
+        (invoke)))
+    (catch com.sun.jdi.InvocationException e
+      (logging/trace
+       "Exception in remote method invocation %s" (.exception e))
+      (throw e))))
 
 ;;; classpath
 (defn classpath