+++ swank/src/ritz/swank/connections.clj
@@ -1,19 +1,19 @@
 (ns ritz.swank.connections)
 
 ;;; # Connections state
-(defonce connections (atom {}))
+(defonce connections (atom #{}))
 
-(defn add-connection [connection proxied-connection]
-  (swap! connections assoc connection proxied-connection))
+(defn add-connection [connection]
+  (swap! connections conj connection))
 
 (defn remove-connection [connection]
-  (swap! connections dissoc connection))
+  (swap! connections disj connection))
 
 (defn connection-for-event
   [_]
-  (ffirst @connections))
+  (first @connections))
 
 (defn all-connections
   "Return all connections."
   []
-  (keys @connections))
+  @connections)