+++ swank/src/ritz/swank/proxy.clj
@@ -6,7 +6,6 @@
    [ritz.jpda.jdi :as jdi]
    [ritz.jpda.jdi-clj :as jdi-clj]
    [ritz.jpda.jdi-vm :as jdi-vm]
-   [ritz.logging :as logging]
    [ritz.repl-utils.compile :as compile]
    [ritz.repl-utils.io :as io]
    [ritz.swank :as swank]
@@ -22,17 +21,21 @@
    ritz.swank.commands.debugger
    ritz.swank.commands.contrib.ritz)
   (:use
-   [ritz.swank.connections :only [add-connection]])
+   [clojure.string :only [join]]
+   [ritz.jpda.jdi :only [invoke-single-threaded]]
+   [ritz.jpda.jdi-clj :only [control-eval]]
+   [ritz.jpda.jdi-vm
+    :only [acquire-thread launch-vm start-control-thread-body vm-resume]]
+   [ritz.logging :only [trace]]
+   [ritz.swank.connections :only [add-connection]]
+   [ritz.swank.rexec :only [rexec rread-msg]]
+   [ritz.swank.exec :only [block-reply-loop]])
   (:import
    com.sun.jdi.VirtualMachine))
 
 (defn forward-commands
   "Alter eval-for-emacs to forward unrecognised commands to proxied connection."
   []
-  ;; (alter-var-root
-  ;;  #'swank/command-not-found
-  ;;  (fn [_ x] x)
-  ;;  debug/forward-command)
   (alter-var-root
    #'swank/forward-rpc
    (fn [_] debug/forward-rpc)))
@@ -45,57 +48,68 @@
       (debug/forward-command
        core/command-not-found))))))
 
+(defn start-remote-thread
+  "Start a remote thread in the specified vm context, using thread-name to
+generate a name for the thread."
+  [vm thread-name]
+  (let [msg-thread-name (name (gensym thread-name))]
+    (acquire-thread
+     vm msg-thread-name
+     (fn [context thread-name]
+       (control-eval
+        context (start-control-thread-body msg-thread-name))))))
+
+;;; # Serve a connection
 (defn serve-connection
   "Serve connection for proxy rpc functions"
   []
-  (logging/trace "proxy/serve-connection")
+  (trace "proxy/serve-connection")
   (.setName (Thread/currentThread) "REPL Proxy")
   (fn proxy-connection-handler
-    [io-connection options]
-    (logging/trace "proxy/proxy-connection-handler")
-    (forward-commands)
+    [io-connection {:keys [log-level classpath extra-classpath vm-classpath]
+                    :as options}]
     (let [options (->
                    options
                    (dissoc :announce)
                    (merge {:port 0 :join true :server-ns 'ritz.repl}))
-          vm-context (debug/launch-vm-with-swank options)
-          options (assoc options :vm-context vm-context)]
-      (logging/trace "proxy/connection-handler: runtime set")
-      (logging/trace "proxy/connection-handler: thread-groups")
-      (logging/trace (with-out-str
-                       (pprint/pprint (jdi/thread-groups (:vm vm-context)))))
+          cp (pr-str (join ":" vm-classpath))
+          vm (apply launch-vm cp `@(promise)
+                    (apply concat (select-keys options [:jvm-opts])))
+          msg-thread (start-remote-thread vm "msg-pump")
+          vm (assoc vm :msg-pump-thread msg-thread)
+          options (assoc options :vm-context vm)]
 
-      (logging/trace "proxy/connection-handler: resume vm")
-      (.resume ^VirtualMachine (:vm vm-context))
+      (debug/add-exception-event-request vm)
+      (trace
+       "proxy/connection-handler exeception events requested")
 
-      (logging/trace "proxy/connection-handler: thread-groups")
-      (logging/trace
-       (with-out-str (pprint/pprint (jdi/thread-groups (:vm vm-context)))))
+      (trace "proxy/connection-handler: resume vm")
+      (vm-resume vm)
 
-      (let [port (debug/remote-swank-port vm-context)]
-        (logging/trace "proxy/connection-handler proxied server on %s" port)
-        (if (= port (:port options))
-          (do
-            (logging/trace "invalid port")
-            ((:close-connection io-connection) io-connection))
-          (let [proxied-connection (debug/create-connection
-                                    (assoc options :port port))
-                _ (logging/trace
-                   "proxy/connection-handler connected to proxied")
-                [connection future] (rpc-server/serve-connection
-                                     io-connection
-                                     (merge
-                                      options
-                                      {:proxy-to proxied-connection
-                                       :swank-handler swank-pipeline}))]
-            (logging/trace "proxy/connection-handler running")
-            (executor/execute-loop
-             (partial debug/forward-reply connection) :name "Reply pump")
-            (logging/trace "proxy/connection-handler reply-pump running")
-            (hooks/run core/new-connection-hook connection)
-            (logging/trace "proxy/connection-handler new-connection-hook ran")
-            (add-connection connection proxied-connection)
-            (logging/trace "proxy/connection-handler connection added")
-            (debug/add-exception-event-request vm-context)
-            (logging/trace
-             "proxy/connection-handler exeception events requested")))))))
+      (ritz.jpda.jdi-clj/control-eval   ; require ritz in the vm
+       vm `(require 'ritz.swank.exec 'ritz.logging))
+      (when log-level                   ; forward the log-level setting
+        (ritz.jpda.jdi-clj/control-eval
+         vm `(ritz.logging/set-level ~log-level)))
+      (trace "proxy/connection-handler: set extra classpath")
+      (ritz.jpda.jdi-clj/control-eval   ; set the extra classpath for ritz
+       vm `(ritz.swank.exec/set-extra-classpath! ~(vec extra-classpath)))
+      (trace "proxy/connection-handler: set classpath")
+      (ritz.jpda.jdi-clj/control-eval   ; set the project classpath
+       vm `(ritz.swank.exec/set-classpath! ~(vec classpath)))
+
+      (trace "proxy/proxy-connection-handler")
+      (forward-commands)
+
+      (let [[connection future] (rpc-server/serve-connection
+                                 io-connection
+                                 (merge
+                                  options
+                                  {:swank-handler swank-pipeline}))]
+        (add-connection connection)
+        (trace "proxy/connection-handler running")
+        (executor/execute-loop
+         (partial debug/forward-reply connection) :name "Reply pump")
+        (trace "proxy/connection-handler reply-pump running")
+        (hooks/run core/new-connection-hook connection)
+        (trace "proxy/connection-handler new-connection-hook ran")))))